Eagle Schematic and Board files
These are the eagle schematic and board files for the flight computer. There are two board files, one with the original Farsi (Persian) silkscreen and another in English with the name in English and a placeholder in place of the slogan. The board layout is identical, and they are only aesthetically different, with the schematic being the same.